Refactoring do programa para registrar todos os número de ponto flutuante como double e para registrar imbalance positivo e negativo. Parametrização do tempo limite de execução do GRASP. Alteração do layout dos arquivos de saída.

+++ mestrado/src/graph/Graph.cpp
@@ -50,7 +50,7 @@ void SignedGraph::addEdge(int a, int b, Edge edge) {
 	add_edge(a, b, edge, graph);
 }
 
-float SignedGraph::getEdge(const int &a, const int &b) {
+double SignedGraph::getEdge(const int &a, const int &b) {
 	return graph.get_edge(a, b).second.weight;
 }
